Se agregó comentarios descriptivos en controllers

+++ src/controllers/comment.controller.js
@@ -1,14 +1,22 @@
 import { CommentModel } from "../models/comment.model.js";
 
+//crear comentario
 export const createComment = async (req, res) => {
     try {
-        const comment = await CommentModel.create(req.body);
+        const { articleId, content } = req.body;
+        const author = req.user._id;
+        const comment = await CommentModel.create({
+            content,
+            author,
+            article: articleId
+        });
         return res.status(201).json({ ok: true, data: comment });
     } catch (error) {
         return res.status(500).json({ ok: false, msg: "Internal server error" });
     }
 };
 
+//obtener todos los comentarios
 export const getAllComments = async (req, res) => {
     try {
         const comments = await CommentModel.find();
@@ -18,6 +26,7 @@ export const getAllComments = async (req, res) => {
     }
 };
 
+//obtener los comentarios por ID
 export const getCommentById = async (req, res) => {
     const { id } = req.params;
     try {
@@ -28,9 +37,14 @@ export const getCommentById = async (req, res) => {
     }
 };
 
+//actualizar comentario
 export const updateComment = async (req, res) => {
     const { id } = req.params;
     try {
+        const exists = await CommentModel.findById(id);
+        if (!exists) {
+            return res.status(404).json({ ok: false, msg: "Comentario no encontrado" });
+        }
         const comment = await CommentModel.findByIdAndUpdate(id, req.body, {
             new: true,
         });
@@ -40,9 +54,14 @@ export const updateComment = async (req, res) => {
     }
 };
 
+//eliminar comentario
 export const deletedComment = async (req, res) => {
     const { id } = req.params;
     try {
+        const exists = await CommentModel.findById(id);
+        if (!exists) {
+            return res.status(404).json({ ok: false, msg: "Comentario no encontrado" });
+        }
         const comment = await CommentModel.findByIdAndDelete(id);
         return res.status(200).json({ ok: true, data: comment });
     } catch (error) {
